Set block reference in ByteBuffer adapters.
According to the contract of the Buffer class, the
effectiveDirectAddress is the address of the data for direct buffers and
null otherwise. It also says that for direct buffers, the block
reference will be set. Otherwise, it's null.

The various <Type>ToByteBuffer adapters wrap ByteBuffers into
<Type>Buffers. While they do so, they set their effectiveDirectAddress
to be equal to the effectiveDirectAddress of the ByteBuffer they are
wrapping. However, they leave the block reference null, which breaks the
contract of the Buffer class.

In this commit, we also set the block reference in the adapters to be
equal to the block reference of the ByteBuffer they are wrapping. As a
result, the "final" attribute of the block reference had to be removed.

This is important for the JNI layer as we need the taint tag in the
MemoryBlock when someone calls GetDirectByteBufferAddress.